requirements.txt
tensorflow
keras
gdown
streamlit
numpy
Pillow
Flask